% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{accumulate}
\alias{accumulate}
\title{array accumulator}
\usage{
accumulate(X, all.dim = NULL, FUN = `+`, na.value = 0, ...)
}
\arguments{
\item{X}{list of arrays. Arrays must have the same number of dimensions (up to 5) but not the same size}

\item{all.dim}{vector of booleans of equal length as the number of dimensions of the arrays}

\item{FUN}{function, reduction function applied element-wise to the matrices}

\item{na.value}{function, value assigned to missing elements}

\item{...}{additional arguments given to FUN}
}
\value{
an array
}
\description{
array accumulator
}
\author{
G.A.Paleologo
}

