% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube.R
\name{add_axes}
\alias{add_axes}
\title{adds one or more dimensions, together with an axis name and a a value of the
coordinate}
\usage{
add_axes(x, ...)
}
\arguments{
\item{x}{array object}

\item{value}{character with variable argument. add_axes(x, height=subject1)
is allowed.}
}
\value{
array object with changed axes names
}
\description{
The dimension is added at the end. The array goes from dimension (m x n x p)
to (m x n x p x 1).
}
\examples{
x2 <- add_axes(Titanic, shipname='titanic')
}

