% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{aggregate_array}
\alias{aggregate_array}
\title{Aggregates an array across the same values of a dimension}
\usage{
aggregate_array(X, FUN, ...)
}
\arguments{
\item{X}{array}

\item{FUN}{function to apply to all elements in the array with same labels
across all axes. In order to obtain reliable results, it is strongly recommended
that the function output be independent of the permutation of its arguments}

\item{...}{additional parameters to pass to FUN}
}
\description{
Aggregates an array across the same values of a dimension
}
\examples{
R <- getTotrets()
axes(R) <- c('date','id')
X <- cuaxismutate(R, da
te=function(x)substr(x, 1,7))
X2 <- aggregate(X, mean)
}

