% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube.R
\name{align_array}
\alias{align_array}
\title{align a list of arrays}
\usage{
align_array(X, all.dim = NULL, na.value = NA)
}
\arguments{
\item{X}{list of arrays. Arrays must have the same number of dimensions (up to 5) but not the same size}

\item{all.dim}{vector of booleans of equal length as the number of dimensions of the arrays}

\item{na.value}{function, value assigned to missing elements}
}
\value{
list of matrices
}
\description{
align a list of arrays
}
\author{
G.A.Paleologo
}

