% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{as.array.data.frame}
\alias{as.array.data.frame}
\title{Obtains an array from a data frame}
\usage{
\method{as.array}{data.frame}(data, FUN, id_var, ...)
}
\arguments{
\item{id_var}{field}

\item{...}{names of columns across which to aggregate}

\item{X}{data frame}
}
\value{
an array
}
\description{
Obtains an array from a data frame
}
\examples{
as.array(iris, sum, "Sepal.Length", "Species", "Petal.Width")
as.array(iris, sum,  Sepal.Length, Species, Petal.Width)
}

