% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube.R
\name{axes<-}
\alias{axes<-}
\title{replacement function that sets axes names for a array. The dimensions
of the array must have names}
\usage{
axes(x) <- value
}
\arguments{
\item{value}{character vector, names of the exes}
}
\value{
object with changed axes names
}
\description{
replacement function that sets axes names for a array. The dimensions
of the array must have names
}
\examples{
A <- array(1:10, dim=c(2,5), dimnames=list(letters[1:2], letters[11:15]))
          axes(A) <- c('height','width')
}

