% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{axis_permute}
\alias{axis_permute}
\title{Permutes dimensions by setting a giving a vector of axes in a different order}
\usage{
axis_permute(X, newaxes)
}
\arguments{
\item{X}{array or matrix}

\item{newaxes}{chararacter vector, axes names of X, but in a different order}
}
\value{
array
}
\description{
Permutes dimensions by setting a giving a vector of axes in a different order
}

