% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{filter_.array}
\alias{filter_.array}
\title{selects a subset of a array in a functional manner}
\usage{
\method{filter_}{array}(X, ...)
}
\arguments{
\item{X}{array}

\item{...}{logical conditions}
}
\value{
a array
}
\description{
selects a subset of a array in a functional manner
}
\examples{
X <- array(1:20, dim=c(4,5), dimnames=list(height=letters[1:4], width=letters[11:15]))
class(X) <- 'array'
filter(X, height >= 'c')
filter(Titanic, Age=='Child', Class \%in\% c('2nd','1st'))
}

