% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{filter_array}
\alias{filter_array}
\title{selects a subset of a array in a functional manner}
\usage{
filter_array(X, ...)
}
\arguments{
\item{X}{array}

\item{...}{logical conditions}
}
\value{
a array
}
\description{
selects a subset of a array in a functional manner
}
\examples{
X <- array(1:20, dim=c(4,5), dimnames=list(height=letters[1:4], width=letters[11:15]))
filter_array(X, height=height >= 'c')
filter_array(Titanic, Age=Age=='Child', Class=Class \%in\% c('2nd','1st'))
}

