% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube.R
\name{find_names}
\alias{find_names}
\title{string matcher}
\usage{
find_names(.data, match_string)
}
\description{
Given a character vector .data and a character vector match_string
returns a character vector of the same length as x.
The i-th element is the element in y that is a substring of
the i-th element of x. If less than one or more than one,
it produces an error. Internal use only
}
\examples{
x <- c('a+b', 'a-d', 'c'); y <- c('b', 'c','d')
find_names(x, y)
x <- c('a+b', 'a-d', 'c'); y <- c('a', 'b', 'c', 'd')
find_names(x, y)  # error: too many matches
x <- c('a+b', 'a-d', 'c'); y <- c('x', 'y', 'z')
find_names(x, y)  # error: no matches
}

