% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{groupby}
\alias{groupby}
\title{Aggregates the individual labels of one or more axes in order to summarize the cell contents}
\usage{
groupby(X, ...)
}
\arguments{
\item{X}{array}

\item{...,}{one or more functions of the form <axisname>= <expression of axisnames>}
}
\value{
an Array
}
\description{
Aggregates the individual labels of one or more axes in order to summarize the cell contents
}

