% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{join_all}
\alias{join_all}
\title{outer joins of two or more n-way arrays, checking they have no dimension names in common}
\usage{
join_all(X)
}
\arguments{
\item{X}{a list of arrays. Arrays must be numeric, integer or logical
and have the same number of dimensions (up to 5) but not necessarily the same
size and their dimnames must have zero intersection}
}
\value{
an array
}
\description{
outer joins of two or more n-way arrays, checking they have no dimension names in common
}
\author{
G.A.Paleologo
}

