% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{notin}
\alias{notin}
\title{returns logical vector if elements of first vector are not in second
vector}
\usage{
notin(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
logical vector
}

