% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{overlay}
\alias{overlay}
\title{Overlays a cube on another}
\usage{
overlay(X, Y)
}
\arguments{
\item{X}{cube}

\item{Y}{cube}
}
\value{
cube
}
\description{
Given two cubes, the function: i) outer aligns them; replaces a NA value in the first one with the
corresponding element of the second
}

