% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube.R
\name{permute_axis}
\alias{permute_axis}
\title{Permutes dimensions by setting a giving a vector of axes in a different order}
\usage{
permute_axis(X, newaxes)
}
\arguments{
\item{X}{array or matrix}

\item{newaxes}{chararacter vector, axes names of X, but in a different order}
}
\value{
array
}
\description{
Permutes dimensions by setting a giving a vector of axes in a different order
}

