% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{rename_axis}
\alias{rename_axis}
\title{Rename one or more axis}
\usage{
rename_axis(X, ...)
}
\arguments{
\item{X}{array or matrix}

\item{...}{named arguments, like oldaxisname = 'newaxisname'}
}
\value{
array
}
\description{
Rename one or more axis
}

