% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{set_axes}
\alias{set_axes}
\title{function that sets axes names, to be used in a pipe}
\usage{
set_axes(x, value)
}
\arguments{
\item{x}{array object}

\item{value}{character vector, names of the exes}
}
\value{
object with changed axes names
}
\description{
function that sets axes names, to be used in a pipe
}

