% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{summary.array}
\alias{summary.array}
\title{A simple array summary}
\usage{
\method{summary}{array}(X)
}
\arguments{
\item{X}{array or matrix}
}
\value{
side effect: print some data about the array
}
\description{
A simple array summary
}

