% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{summary_raggedarray}
\alias{summary_raggedarray}
\title{This function prints summary statistics of a ragged array
i.e. a list f vectors of unequal length. Mostly used with
lists of portfolios}
\usage{
summary_raggedarray(X)
}
\arguments{
\item{X}{list, a list comprised of vectors coercible to numeric}
}
\value{
the function has only side effects
}
\description{
This function prints summary statistics of a ragged array
i.e. a list f vectors of unequal length. Mostly used with
lists of portfolios
}

