% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{timeit}
\alias{timeit}
\title{a simple decorator to benchmark execution times}
\usage{
timeit(f)
}
\arguments{
\item{f}{function}
}
\value{
a function. Identical to f, but prints execution time
}
\description{
a simple decorator to benchmark execution times
}

