% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube.R
\name{df2c}
\alias{df2c}
\title{Obtains an array from a data frame}
\usage{
df2c(data, FUN, id_var, ...)
}
\arguments{
\item{id_var}{field}

\item{...}{names of columns across which to aggregate}

\item{X}{data frame}
}
\value{
an array
}
\description{
Obtains an array from a data frame
}
\examples{
df2c(iris, "Sepal.Length", "Species", "Petal.Width")
          df2c(iris, Sepal.Length, Species, Petal.Width)
}

