% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{overlay}
\alias{overlay}
\title{Overlays a array on another}
\usage{
overlay(X, Y)
}
\arguments{
\item{X}{array}

\item{Y}{array}
}
\value{
array
}
\description{
Given two arrays, the function: i) outer aligns them; replaces a NA value in the first one with the 
corresponding element of the second
}

